# backend/core/policy_engine/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Grandway Policy-Registered API",
    "version": "1.0.0",
    "description": "Generated from the Core Policy Engine registry (source of truth for endpoints, permission keys, HTTP methods, and risk levels). Request/response body schemas are documented per app in `docs/DATA_CONTRACT.md`."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/v1/applicants/": {
      "get": {
        "operationId": "applicants.applicant.list",
        "summary": "List Applicants",
        "description": "List every applicant. Shared across the consultancy — not owner-scoped. Searchable by name, email, contact number, and passport number, and filterable by destination country and journey stage.\n\nRequest/response field schemas: see the owning app's `docs/DATA_CONTRACT.md`.",
        "tags": [
          "applicants"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission-key": "applicants.applicant.list",
        "x-risk-level": "low",
        "x-operation-type": "list",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "operationId": "applicants.applicant.create",
        "summary": "Create Applicant",
        "description": "Create an applicant directly, with no preceding lead. Admin only.\n\nRequest/response field schemas: see the owning app's `docs/DATA_CONTRACT.md`.",
        "tags": [
          "applicants"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission-key": "applicants.applicant.create",
        "x-risk-level": "high",
        "x-operation-type": "create",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/applicants/{applicant_id}/": {
      "get": {
        "operationId": "applicants.applicant.read",
        "summary": "View Applicant",
        "description": "Retrieve one applicant with contacts, addresses, passport, family, and emergency contacts.\n\nRequest/response field schemas: see the owning app's `docs/DATA_CONTRACT.md`.",
        "tags": [
          "applicants"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission-key": "applicants.applicant.read",
        "x-risk-level": "low",
        "x-operation-type": "read",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "applicant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "patch": {
        "operationId": "applicants.applicant.update",
        "summary": "Edit Applicant",
        "description": "Correct identity, contact, address, passport, family, or emergency-contact information.\n\nRequest/response field schemas: see the owning app's `docs/DATA_CONTRACT.md`.",
        "tags": [
          "applicants"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission-key": "applicants.applicant.update",
        "x-risk-level": "medium",
        "x-operation-type": "update",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "applicant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/journeys/": {
      "get": {
        "operationId": "applicant_journeys.journey.list",
        "summary": "List Journeys",
        "description": "List study objectives across all applicants. Shared, not owner-scoped.\n\nRequest/response field schemas: see the owning app's `docs/DATA_CONTRACT.md`.",
        "tags": [
          "applicant_journeys"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission-key": "applicant_journeys.journey.list",
        "x-risk-level": "low",
        "x-operation-type": "list",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "operationId": "applicant_journeys.journey.create",
        "summary": "Create Journey",
        "description": "Record a new overseas-study objective for an existing applicant.\n\nRequest/response field schemas: see the owning app's `docs/DATA_CONTRACT.md`.",
        "tags": [
          "applicant_journeys"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission-key": "applicant_journeys.journey.create",
        "x-risk-level": "medium",
        "x-operation-type": "create",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/journeys/{journey_id}/": {
      "get": {
        "operationId": "applicant_journeys.journey.read",
        "summary": "View Journey",
        "description": "Retrieve one journey with its objective details and lifecycle state.\n\nRequest/response field schemas: see the owning app's `docs/DATA_CONTRACT.md`.",
        "tags": [
          "applicant_journeys"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission-key": "applicant_journeys.journey.read",
        "x-risk-level": "low",
        "x-operation-type": "read",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "journey_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "patch": {
        "operationId": "applicant_journeys.journey.update",
        "summary": "Edit Journey",
        "description": "Correct destination, level, field, intake, budget, or notes. Stage is unaffected.\n\nRequest/response field schemas: see the owning app's `docs/DATA_CONTRACT.md`.",
        "tags": [
          "applicant_journeys"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission-key": "applicant_journeys.journey.update",
        "x-risk-level": "medium",
        "x-operation-type": "update",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "journey_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/journeys/{journey_id}/stage/": {
      "post": {
        "operationId": "applicant_journeys.journey.change_stage",
        "summary": "Change Journey Stage",
        "description": "Move a journey between active stages. Cannot set completed, closed, or deferred.\n\nRequest/response field schemas: see the owning app's `docs/DATA_CONTRACT.md`.",
        "tags": [
          "applicant_journeys"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission-key": "applicant_journeys.journey.change_stage",
        "x-risk-level": "medium",
        "x-operation-type": "custom",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "journey_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/leads/{lead_id}/stage/": {
      "post": {
        "operationId": "leads.lead.change_stage",
        "summary": "Change Lead Stage",
        "description": "Move a lead between active stages. Cannot set lost or converted.\n\nRequest/response field schemas: see the owning app's `docs/DATA_CONTRACT.md`.",
        "tags": [
          "leads"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission-key": "leads.lead.change_stage",
        "x-risk-level": "medium",
        "x-operation-type": "custom",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/policy/apps/": {
      "get": {
        "operationId": "policy_engine.application.list",
        "summary": "List Policy Applications",
        "description": "List all registered policy applications in the Core Policy Engine.\n\nRequest/response field schemas: see the owning app's `docs/DATA_CONTRACT.md`.",
        "tags": [
          "policy_engine"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission-key": "policy_engine.application.list",
        "x-risk-level": "low",
        "x-operation-type": "list",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/policy/apps/{app_key}/": {
      "get": {
        "operationId": "policy_engine.application.read",
        "summary": "View Policy Application Detail",
        "description": "Retrieve a single policy application with its models and endpoints.\n\nRequest/response field schemas: see the owning app's `docs/DATA_CONTRACT.md`.",
        "tags": [
          "policy_engine"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission-key": "policy_engine.application.read",
        "x-risk-level": "low",
        "x-operation-type": "read",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "app_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "SuccessEnvelope": {
        "type": "object",
        "required": [
          "success",
          "message",
          "data"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "message": {
            "type": "string"
          },
          "data": {
            "description": "Resource payload — see the owning app's DATA_CONTRACT.md."
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "success",
          "error"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "APP_RESOURCE_REASON code."
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object"
              }
            }
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        }
      },
      "Meta": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "page_size": {
            "type": "integer"
          },
          "next": {
            "type": [
              "string",
              "null"
            ]
          },
          "previous": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      }
    },
    "responses": {
      "Error": {
        "description": "Standard error envelope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    }
  }
}